Refactoring for modules
