+++ .idea/gradle.xml
@@ -9,6 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/capital-gain-adapter-cli" />
+            <option value="$PROJECT_DIR$/capital-gain-core" />
           </set>
         </option>
       </GradleProjectSettings>
