+++ capital-gain-adapter-cli/src/main/java/com/nubank/capitalgain/adapter/cli/CapitalGainMain.java
@@ -1,0 +1,11 @@
+package com.nubank.capitalgain.adapter.cli;
+
+import com.nubank.capitalgain.adapter.cli.in.TaxCalculatorAdapterCli;
+
+public class CapitalGainMain {
+
+    public static void main(String[] args) {
+        TaxCalculatorAdapterCli taxCalculatorAdapterConsoleCli = new TaxCalculatorAdapterCli();
+        taxCalculatorAdapterConsoleCli.calculate();
+    }
+}
