+++ capital-gain-adapter-cli/src/main/java/com/nubank/capitalgain/adapter/cli/in/JsonMapper.java
@@ -1,0 +1,31 @@
+package com.nubank.capitalgain.adapter.cli.in;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.MapperFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+class JsonMapper {
+
+    private final ObjectMapper objectMapper;
+
+    public JsonMapper() {
+        this.objectMapper = new ObjectMapper();
+        objectMapper.configure(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS, true);
+    }
+
+    public String writeValuesAsString(Object object) {
+        try {
+            return objectMapper.writeValueAsString(object);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException("Error serializing object to JSON", e);
+        }
+    }
+
+    public Object readValue(String json, Class<?> valueType) {
+        try {
+            return objectMapper.readValue(json, valueType);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException("Error deserializing JSON to object", e);
+        }
+    }
+}
