+++ capital-gain-adapter-cli/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module com.nubank.capitalgain.adapter.cli {
+    requires com.fasterxml.jackson.core;
+    requires com.fasterxml.jackson.databind;
+    requires com.nubank.capitalgain.core.module;
+
+    exports com.nubank.capitalgain.adapter.cli;
+}
