+++ capital-gain-core/src/main/java/com/nubank/capitalgain/core/calculator/application/dto/FeeDto.java
@@ -1,0 +1,10 @@
+package com.nubank.capitalgain.core.calculator.application.dto;
+
+import java.math.BigDecimal;
+
+public record FeeDto(BigDecimal tax) {
+
+    public FeeDto(String tax) {
+        this(new BigDecimal(tax));
+    }
+}
