+++ capital-gain-core/src/main/java/com/nubank/capitalgain/core/calculator/application/dto/TradeDto.java
@@ -1,0 +1,11 @@
+package com.nubank.capitalgain.core.calculator.application.dto;
+
+import com.nubank.capitalgain.core.calculator.model.TypeOperation;
+import java.math.BigDecimal;
+
+public record TradeDto(TypeOperation typeOperation, BigDecimal unitCost, int quantity) {
+
+    public TradeDto(TypeOperation typeOperation, String unitCost, int quantity) {
+        this(typeOperation, new BigDecimal(unitCost), quantity);
+    }
+}
