+++ capital-gain-core/src/main/java/com/nubank/capitalgain/core/calculator/domain/Calculator.java
@@ -1,0 +1,18 @@
+package com.nubank.capitalgain.core.calculator.domain;
+
+/**
+ * Interface to calculate a value based on a given input.
+ *
+ * @param <T> the input typeOperation
+ * @param <R> the output typeOperation
+ */
+public interface Calculator<T, R> {
+
+    /**
+     * Compute a value based on the given input.
+     *
+     * @param t the input
+     * @return the computed value
+     */
+    R computedValue(T t);
+}
