+++ capital-gain-core/src/main/java/com/nubank/capitalgain/core/calculator/exceptions/IllegalOperationStockException.java
@@ -1,0 +1,8 @@
+package com.nubank.capitalgain.core.calculator.exceptions;
+
+public class IllegalOperationStockException extends RuntimeException {
+
+    public IllegalOperationStockException(String message) {
+        super(message);
+    }
+}
