+++ capital-gain-core/src/main/java/com/nubank/capitalgain/core/calculator/model/TypeOperation.java
@@ -1,0 +1,10 @@
+package com.nubank.capitalgain.core.calculator.model;
+
+/**
+ * Enum for the typeOperation of operation.
+ */
+public enum TypeOperation {
+
+    BUY,
+    SELL
+}
