+++ capital-gain-core/src/main/java/com/nubank/capitalgain/core/commons/ZeroMoney.java
@@ -1,0 +1,10 @@
+package com.nubank.capitalgain.core.commons;
+
+import java.math.BigDecimal;
+
+public final class ZeroMoney extends Money {
+
+    public ZeroMoney() {
+        super(BigDecimal.ZERO);
+    }
+}
