+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'capital-gain'
-
+include 'capital-gain-core'
+include 'capital-gain-adapter-cli'
