--- src/main/java/com/nubank/capitalgain/CapitalGainMain.java
@@ -1,11 +1,0 @@
-package com.nubank.capitalgain;
-
-import com.nubank.capitalgain.adapter.in.TaxCalculatorAdapterConsoleCli;
-
-public class CapitalGainMain {
-
-    public static void main(String[] args) {
-        TaxCalculatorAdapterConsoleCli taxCalculatorAdapterConsoleCli = new TaxCalculatorAdapterConsoleCli();
-        taxCalculatorAdapterConsoleCli.calculate();
-    }
-}
