--- src/main/java/com/nubank/capitalgain/application/domain/commons/TypeOperation.java
@@ -1,10 +1,0 @@
-package com.nubank.capitalgain.application.domain.commons;
-
-/**
- * Enum for the typeOperation of operation.
- */
-public enum TypeOperation {
-
-    BUY,
-    SELL
-}
