--- src/main/java/com/nubank/capitalgain/application/domain/exceptions/IllegalOperationStockException.java
@@ -1,8 +1,0 @@
-package com.nubank.capitalgain.application.domain.exceptions;
-
-public class IllegalOperationStockException extends RuntimeException {
-
-    public IllegalOperationStockException(String message) {
-        super(message);
-    }
-}
