--- src/main/java/com/nubank/capitalgain/application/domain/primitives/ZeroMoney.java
@@ -1,10 +1,0 @@
-package com.nubank.capitalgain.application.domain.primitives;
-
-import java.math.BigDecimal;
-
-public final class ZeroMoney extends Money {
-
-    public ZeroMoney() {
-        super(BigDecimal.ZERO);
-    }
-}
